chore: remove legacy footer leftovers
