+++ QA_CHECKLIST.md
@@ -1,0 +1,9 @@
+# QA Manual Checklist
+
+- [ ] 375×667 mobile: home, shop, product, cart → checkout → success/failure
+- [ ] 768×1024 tablet: home, shop, product, cart → checkout → success/failure
+- [ ] 1280px desktop: home, shop, product, cart → checkout → success/failure
+- [ ] Admin panel footer renders with CTA hidden
+- [ ] Tab through footer links, focus visible and keyboard navigation works
+- [ ] Contact chips, legal links and social icons resolve correctly
+- [ ] WhatsApp FAB offset adjusts when footer or fixed elements appear
