+++ nerin_final_updated/frontend/components/np-footer.css
@@ -5,7 +5,7 @@
   --color-text: #1a1a1a;
   --color-border: rgba(0, 0, 0, 0.08);
   --cta-h: 0px;
-  --wa-bottom: 16px;
+  --fab-safe-offset: 0px;
 }
 
 /* Sticky CTA */
@@ -23,7 +23,7 @@
   justify-content: space-between;
   gap: var(--space-12);
   min-height: 56px;
-  z-index: 960;
+  z-index: 300;
   transform: translateZ(0);
   transition: transform 0.18s ease, opacity 0.18s ease;
   will-change: transform;
@@ -44,14 +44,14 @@
   width: 56px;
   height: 56px;
   right: 16px;
-  bottom: calc(var(--wa-bottom) + env(safe-area-inset-bottom));
+  bottom: calc(16px + var(--fab-safe-offset) + env(safe-area-inset-bottom));
   border-radius: 50%;
   background: #25d366;
   display: flex;
   align-items: center;
   justify-content: center;
   text-decoration: none;
-  z-index: 990;
+  z-index: 310;
 }
 
 .wa-fab:focus-visible {
@@ -63,22 +63,19 @@
   transition: transform 0.18s ease, opacity 0.18s ease;
 }
 
-/* opción A: ocultar WA cuando el footer está visible */
+/* hide WhatsApp FAB when footer is visible */
 html.footer-visible [data-wa] {
   opacity: 0;
   pointer-events: none;
   transform: translateY(20%);
 }
 
-/* opción B (si NO querés ocultarlo): subirlo 96px cuando el footer entra */
-html.footer-visible.move-wa [data-wa] {
-  opacity: 1;
-  pointer-events: auto;
-  transform: translateY(-96px);
-}
-
 main {
   padding-bottom: calc(var(--cta-h) + env(safe-area-inset-bottom));
+}
+
+body {
+  padding-bottom: calc(var(--fab-safe-offset) + 64px);
 }
 
 @media (max-width: 360px) {
@@ -87,10 +84,6 @@ main {
     height: 52px;
     right: 12px;
   }
-}
-
-html.cta-visible .wa-fab {
-  bottom: calc(var(--wa-bottom) + var(--cta-h) + env(safe-area-inset-bottom));
 }
 
 body.hide-cta [data-sticky-cta] {
@@ -141,11 +134,8 @@ body.hide-cta main {
   outline: 2px auto;
 }
 
-.footer-nav .footer-links,
+.footer-nav,
 .footer-social {
-  list-style: none;
-  margin: 0;
-  padding: 0;
   display: flex;
   flex-wrap: wrap;
   gap: var(--space-12);
