+++ nerin_final_updated/frontend/components/np-footer.html
@@ -12,18 +12,16 @@
     </div>
 
     <nav class="footer-nav" aria-label="Enlaces">
-      <ul class="footer-links">
-        <li><a href="/index.html">Inicio</a></li>
-        <li><a href="/shop.html">Catálogo</a></li>
-        <li><a href="/contact.html">Contacto</a></li>
-        <li><a href="/seguimiento.html">Seguir mi pedido</a></li>
-      </ul>
+      <a href="/index.html">Inicio</a>
+      <a href="/shop.html">Catálogo</a>
+      <a href="/contact.html">Contacto</a>
+      <a href="/seguimiento.html">Seguir mi pedido</a>
     </nav>
 
-    <ul class="footer-social">
-      <li><a href="https://instagram.com/nerinparts" target="_blank" rel="noopener" aria-label="Instagram">Instagram</a></li>
-      <li><a href="https://linkedin.com/company/nerinparts" target="_blank" rel="noopener" aria-label="LinkedIn">LinkedIn</a></li>
-    </ul>
+    <div class="footer-social">
+      <a href="https://instagram.com/nerinparts" target="_blank" rel="noopener" aria-label="Instagram">Instagram</a>
+      <a href="https://linkedin.com/company/nerinparts" target="_blank" rel="noopener" aria-label="LinkedIn">LinkedIn</a>
+    </div>
 
     <p class="legal">
       © 2025 NERIN PARTS — CUIT XX-XXXXXXXX-X — Exento/Convenio — <a href="/terminos.html">Términos</a> — <a href="/privacidad.html">Privacidad</a>
