--- nerin_final_updated/frontend/css/footer.css
@@ -1,171 +1,0 @@
-:root {
-  --brand: #0A5CFF;
-  --ink: #0E1116;
-  --muted: #5B6676;
-  --line: #E7EBF0;
-  --bg: #FFF;
-  --bg-soft: #F6F8FA;
-  --radius: 10px;
-  --container: 1200px;
-  --fab-safe-offset: 0px;
-  --space-8: 8px;
-  --space-12: 12px;
-  --space-16: 16px;
-}
-
-.site-footer {
-  border-top: 1px solid var(--line);
-  background: var(--bg);
-  color: var(--ink);
-  font-size: clamp(14px, 1.6vw, 16px);
-  line-height: 1.5;
-}
-
-.site-footer .container {
-  max-width: var(--container);
-  margin-inline: auto;
-  padding: var(--space-12) 16px;
-  padding-bottom: calc(var(--space-12) + env(safe-area-inset-bottom));
-}
-
-.footer-grid {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-12);
-}
-
-.contact-chips .chips {
-  display: flex;
-  flex-wrap: wrap;
-  gap: var(--space-8);
-}
-
-.chip {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--space-8);
-  padding: 10px 12px;
-  border: 1px solid var(--line);
-  border-radius: 999px;
-  background: var(--bg-soft);
-  color: var(--ink);
-  text-decoration: none;
-  min-height: 44px;
-}
-
-.chip:hover,
-.chip:focus-visible {
-  text-decoration: underline;
-  outline: 2px auto;
-}
-
-.footer-nav,
-.footer-social {
-  display: flex;
-  flex-wrap: wrap;
-  gap: var(--space-12);
-}
-
-
-.footer-nav a,
-.footer-social a {
-  display: flex;
-  align-items: center;
-  min-height: 44px;
-  color: var(--ink);
-  text-decoration: none;
-}
-
-.footer-social a {
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  min-height: 32px;
-  border: 1px solid var(--line);
-  border-radius: 8px;
-  background: var(--bg-soft);
-  color: var(--muted);
-}
-
-.footer-social a svg {
-  width: 18px;
-  height: 18px;
-  fill: currentColor;
-}
-
-.footer-nav a:hover,
-.footer-nav a:focus-visible,
-.footer-social a:hover,
-.footer-social a:focus-visible {
-  color: var(--brand);
-  text-decoration: underline;
-  outline: 2px auto;
-}
-
-.footer-social a:hover,
-.footer-social a:focus-visible {
-  text-decoration: none;
-}
-
-.legal {
-  color: var(--muted);
-  font-size: clamp(12px, 1.6vw, 13.5px);
-  max-width: 70ch;
-}
-
-.legal a {
-  color: inherit;
-  text-decoration: underline;
-}
-
-@media (min-width: 600px) and (max-width: 1024px) {
-  .site-footer .container {
-    padding: var(--space-16);
-    padding-bottom: calc(var(--space-16) + env(safe-area-inset-bottom));
-  }
-
-  .footer-grid {
-    display: grid;
-    gap: var(--space-16);
-    grid-template-columns: 1fr 1fr;
-    grid-template-areas: "chips nav" "social legal";
-  }
-
-  .contact-chips { grid-area: chips; }
-  .footer-nav { grid-area: nav; }
-  .footer-social { grid-area: social; }
-  .legal { grid-area: legal; }
-
-  .chip,
-  .footer-nav a,
-  .footer-social a {
-    min-height: 40px;
-  }
-}
-
-@media (min-width: 1024px) {
-  .site-footer .container {
-    padding: 24px;
-    padding-bottom: calc(24px + env(safe-area-inset-bottom));
-  }
-
-  .footer-grid {
-    display: grid;
-    gap: 16px;
-    grid-template-columns: 2fr 1fr 1fr 1fr;
-  }
-}
-
-body {
-  padding-bottom: calc(var(--fab-safe-offset) + 64px);
-}
-
-[class*="whatsapp"],
-a[href*="wa.me"],
-a[href*="whatsapp"],
-.whatsapp-fab {
-  position: fixed;
-  right: 16px;
-  bottom: calc(16px + var(--fab-safe-offset));
-  z-index: 1000;
-}
