+++ nerin_final_updated/frontend/seguimiento.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Seguimiento de tu pedido – NERIN</title>
           <link rel="stylesheet" href="style.css?v=mobfix-1" />
-    <link rel="stylesheet" href="/css/footer.css?v=nerin-20250826-1459">
+    <link rel="stylesheet" href="/components/np-footer.css">
   <link rel="stylesheet" href="/css/overrides.css?v=nerin-20250826-1459">
 </head>
   <body>
@@ -51,43 +51,10 @@
         <a href="/index.html" class="button secondary">Volver al inicio</a>
       </div>
     </main>
-<footer class="site-footer">
-  <div class="container footer-grid">
-    <section class="contact-chips">
-      <div class="chips">
-        <a class="chip" href="tel:+5491100000000" aria-label="Llamar por teléfono">
-          <i class="icon-phone"></i> +54 9 11 0000-0000
-        </a>
-        <a class="chip" href="mailto:ventas@nerinparts.com.ar" aria-label="Enviar email">
-          <i class="icon-mail"></i> ventas@nerinparts.com.ar
-        </a>
-        <span class="chip" aria-label="Ubicación">📍 CABA</span>
-      </div>
-    </section>
-
-    <nav class="footer-nav" aria-label="Navegación del sitio">
-      <a href="/index.html">Inicio</a>
-      <a href="/shop.html">Catálogo</a>
-      <a href="/contact.html">Contacto</a>
-      <a href="/seguimiento.html">Seguir mi pedido</a>
-    </nav>
-
-    <section class="footer-social" aria-label="Redes sociales">
-      <a href="https://instagram.com/nerinparts" target="_blank" rel="noopener">Instagram</a>
-      <a href="https://linkedin.com/company/nerinparts" target="_blank" rel="noopener">LinkedIn</a>
-    </section>
-
-    <section class="legal">
-      <div>© 2025 NERIN PARTS — CUIT XX-XXXXXXXX-X — Exento/Convenio</div>
-      <div><a href="/terminos.html">Términos</a> — <a href="/privacidad.html">Privacidad</a></div>
-      <div><small>nerinparts.com.ar</small></div>
-    </section>
-  </div>
-</footer>
 
     <script type="module" src="/js/seguimiento.js"></script>
     <script type="module" src="/js/config.js"></script>
-    <script defer src="/js/footer.js"></script>
-  <script defer src="/js/fab-offset.js"></script>
+  <script defer src="/components/np-footer.js"></script>
+<script defer src="/js/fab-offset.js"></script>
 </body>
 </html>
